fixed some erros

+++ src/main/java/com/example/ecommerce/models/OrderItem.java
@@ -23,6 +23,6 @@ public class OrderItem {
     private int quantity;
 
     @Column(precision = 10, scale = 2)
-    private long price;
+    private BigDecimal price;
 
 }
